Add seabios and qemu dependencies version
Seabios require at least version 1.7.4-2 for 256k build, setted 1.7.4-4
that not have critical or important bugs FWIK.
Qemu version with no critical or important bugs FWIK is >=1.6.1 but >= 1.7
is needed for support seabios 256k build.

This change is useful mainly for backports new xen packages to wheezy and
and have a good hvm domUs support.